corrected typecasting of returned pjump for mu_i and sigma2_i proposals

+++ src/stree.c
@@ -4431,7 +4431,7 @@ double prop_locusrate_sigma2i(gtree_t ** gtree,
       }
     }
   }
-  return accepted / opt_locus_count;
+  return ((double)accepted / opt_locus_count);
 }
 
 double prop_locusrate_mui(gtree_t ** gtree,
@@ -4606,7 +4606,7 @@ double prop_locusrate_mui(gtree_t ** gtree,
       }
     }
   }
-  return accepted / opt_locus_count;
+  return ((double)accepted / opt_locus_count);
 }
 
 long prop_locusrate_mubar(stree_t * stree, gtree_t ** gtree)
